fix: fix success bug when deleting the same id for pets

+++ src/app/service/PetService.js
@@ -11,6 +11,10 @@ module.exports = class PetService {
 
     static async updatePet(petId, petData) {
         try {
+            const pet = await PetService.getByPetId(petId)
+            if (!pet) {
+                throw new Error('Pet not found');
+            }
             await PetRepository.updatePet(petId, petData); 
         } catch (error) {
             throw new Error(error.message);
@@ -19,9 +23,21 @@ module.exports = class PetService {
 
     static async removePet(petId) {
         try {
+            const pet = await PetService.getByPetId(petId)
+            if (!pet) {
+                throw new Error('Pet not found');
+            }
             await PetRepository.removePet(petId)
         } catch (error) {
             throw new Error(error.message);
         }  
     }
-}; 
+
+    static async getByPetId(petId) {
+        try {
+            return await PetRepository.getByPetId(petId)
+        } catch (error) {
+            throw new Error(error.message);
+        }
+    }
+};
